Gets basic canvas rendering text
