--- src/components/image.tsx
@@ -1,12 +1,0 @@
-/** @jsx jsx */
-import { jsx } from '@emotion/core'
-import Header from "./header"
-import Form from "./form"
-
-export default function(){
-  return (
-    <div>
-      Image
-    </div>
-  )
-}
