--- src/form.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="stylesheet" href="main.css" />
-    <title>Editions Card Builder</title>
-  </head>
-
-  <body>
-    <div class="card-builder">
-
-      <div class="card-builder-right"></div>
-    </div>
-    <script src="main.ts" type="text/javascript"></script>
-  </body>
-</html>
