+++ src/index.tsx
@@ -1,7 +1,6 @@
-
 /** @jsx jsx */
 import { jsx } from '@emotion/core'
 import { render } from "react-dom";
-import app from "./components/app"
+import App from "./components/app"
 
-render(app(), document.getElementById('root'));
+render(<App />, document.getElementById('root'));
