+++ src/types/furniture.d.ts
@@ -1,0 +1,10 @@
+export interface Furniture {
+  device: any
+  imageUrl: any
+  headline: any
+  headlineSize: any
+  standfirst: any
+  standfirstSize: any
+  colourCode: any
+  position: any
+}
